+++ packages/globalpayments-js/src/ui/iframe-field/action-add-google-pay.ts
@@ -337,6 +337,7 @@ export default function addGooglePay(iframeField: IframeField | undefined, field
         country: paymentAddress.countryCode,
         phone: paymentAddress.phoneNumber,
         postalCode: paymentAddress.postalCode,
+        name: paymentAddress.name
       }
       return paymentShippingAddress;
     } else return '';